Refactoring and optimized for discrete graphics
Before this update, the simulation ran slow on high performance machines
that had discrete graphics but lacked integrated graphics. It ran really
fast on all systems that had integrated graphics card, but it did not
take advantage of the GPU. This implementation of VolatileImage takes
advantage of the GPU and now the program should run fast provided there
is sufficient ram and a GPU. Code refactoring and comments are added. My
use of VolatileImage does provide a huge performance boost but the
implementation is incomplete.