Implemented button functionality. Code refactored.
User can now Play, Pause, Resume, Stop, and watch a replay of the
previous simulation. User can also now set simulation speed by factors
of 0.5, 1.0, 1.5, and 2.0. The only remaining button to implement is the
map editor button which allows the user to generate a new map or build a
map from scratch.