Implementation of ReplaySlider.
ReplaySlider is a bar at the bottom of map that is available in Replay
Mode. This bar is like a video progress bar that allows a user to "seek"
or jump to a point in the video. The ReplaySlider is a bar that allows
the user to jump to a specific point in the replay.