Major improvements and bug fixes
GUI improvements: more TREE_FACTOR_SIZE variables are added for smaller
factors. Improved the user interface with incomplete menu bar and full
screen view. Added a circular mouse cursor to show the user which trees
will ignite on click.

Bug fixes: Click drag now ignites trees (before trees ignite on single
click but not on drag). Clicking in an area causes nearby trees to burn
in a much more accurate way (before this it was very buggy).

Notice: the new code is less efficient and much slower. Larger tree
factors are very laggy. I'm looking for a solution to make it handle
large amounts of trees like it used to. Still, the performance is not
bad.