Information added, worked on the GUI .
Added copyright and class description. Worked on and improved the GUI
design for the "Create new map" JDialog. Used a JSpinner instead of
JComboBox to choose tree population. Not sure which one I want to keep
so the JComboBox version has been commented out (instead of deleted).
Refactored container variable names. Added mouse functionality for
JDialogs. Menu > Help > About now opens a JDialog that contains the html
and css page that provides the information about the wildfire simulator
and the developer.