Improvements
Improved playback behavior. User can now resume playback if map is
clicked while in pause mode.

Improved look of application. Pause darkens the map and draws the String
"PAUSED" at center of the map.

The JFormatedTextField from the JSpinner in the Configure New Map dialog
can now be right clicked and left clicked. Left click increments the
population factor while right click decrements the population factor.
The clicking will cycle through the list if the spinner reaches the
beginning or end of the list (depending on increment or decrement).

Created the play() method to reduce redundancy.

Fixed a bug that causes array index out of bounds exception (by changing
some numTrees to sortedTrees.size() in ForestFire). Prior to this fix,
clicking on Apply in the "Configure New Map" dialog would trigger the
exception when the user chooses a population factor that is larger than
the current. This is because paintComponent would try to access the
sortedTrees elements at an index greater than sortedTrees.size()
(obvious). numTrees becomes greater than sortedTree.size(), which
triggers the exception in the for  loop. Also, if the user chose a
factor smaller than the current factor, then paintComponent would paint
only a fraction of the sortedTrees. This is because sortedTrees had not
yet been cleared and numTrees has decreased.