Many changes (read description)
Edge scrolling with mouse added--created another Timer to implement
this. Mouse drag scrolling added.

Removed unnecessary print statements. Refactored code in various
classes. Renamed variables in various classes (such as changing width to
mapWidth and jmiPositions to jcbmiPositions). Restructured nested
if-else statements to work more efficiently (for mouse event handling).

Removed some map size options. Worked on sbApply so that Map Size
options now changes the size of the map on Apply.

Added (sortedTrees.size() > 0)  in clickMakeLand method to prevent some
exceptions from happening. I recall it was an index out of bounds
exception.

Created method setMapSize for setting the dimensions of a map.

Created GameStats class which is a JWindow that shows game or map
information at the corner of the map. Several methods are related to
this class: updateGameStats(), hideGameStats(boolean), and
setGameStatsVisible(boolean). hideGameStats is used to hide the
gameStats by putting in behind the JFrame. setGameStatsVisible is used
to hide the gameStats by making the it completely invisible. Both
methods to hide gameStats are necessary. Added MenuListeners to
temporarily hide GameStats window because it is always on top (even when
a menu from the menu bar is expanded). GameStats stays above the
expanded menu, but MenuListeners hide GameStats when the user is using
the menus. Added WindowListeners to also temporarily hide GameStats
window when the user alt-tabs.

LookAndFeel updated with alpha values and new Colors.